Bug 994 On delete if the index exists delete it don't also check for the job doc in the index
Former-commit-id: 2b2020274d554c4bf1b57d662884853b6c5019d9

Original commit: elastic/x-pack@c15b8197bc8bee0e8856a48930d270ee1c4c048d

+++ java/apps/engineApi/engine-api/src/main/java/com/prelert/job/persistence/elasticsearch/ElasticsearchJobProvider.java
@@ -248,15 +248,20 @@ public class ElasticsearchJobProvider implements JobProvider
     @Override
     public boolean jobIdIsUnique(String jobId) throws JobIdAlreadyExistsException
     {
-        IndicesExistsResponse res =
-                m_Client.admin().indices().exists(new IndicesExistsRequest(jobId)).actionGet();
-
-        if (res.isExists())
+        if (indexExists(jobId))
         {
             throw new JobIdAlreadyExistsException(jobId);
         }
 
         return true;
+    }
+
+    private boolean indexExists(String jobId)
+    {
+        IndicesExistsResponse res =
+                m_Client.admin().indices().exists(new IndicesExistsRequest(jobId)).actionGet();
+
+        return res.isExists();
     }
 
 
@@ -501,10 +506,17 @@ public class ElasticsearchJobProvider implements JobProvider
     {
         try
         {
-            DeleteIndexResponse response = m_Client.admin()
-                    .indices().delete(new DeleteIndexRequest(jobId)).get();
+            if (indexExists(jobId))
+            {
+                DeleteIndexResponse response = m_Client.admin()
+                        .indices().delete(new DeleteIndexRequest(jobId)).get();
 
-            return response.isAcknowledged();
+                return response.isAcknowledged();
+            }
+            else
+            {
+                throw new UnknownJobException(jobId);
+            }
         }
         catch (InterruptedException|ExecutionException e)
         {
